requirements.txt
anyio==4.10.0
APScheduler==3.11.0
asgiref==3.9.1
asttokens==3.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
executing==2.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
mysqlclient==2.2.7
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.13
